+++ processor/src/main/java/com/anggrayudi/materialpreference/processor/PreferenceSpec.kt
@@ -1,0 +1,9 @@
+package com.anggrayudi.materialpreference.processor
+
+data class PreferenceSpec(
+        var key: String = "",
+        var constantName: String = "",
+        var methodName: String? = null,
+        var dataType: PreferenceDataType = PreferenceDataType.NOTHING,
+        var defaultValue: Any? = null
+)