+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 
+## 3.5.0 (2019-09-02)
+
+### Enhancements
+* Annotation processor will generate `SharedPreferencesHelper`, so you can don't need to get `SharedPreferences` value like this: `SharedPreferences.get<DataType>(key, defaultValue)`.
+Take advantage of using it with dependency injection such as [Dagger 2](https://github.com/google/dagger) and [Koin](https://github.com/InsertKoinIO/koin).
+Personally, I would recommend you to use Koin because of its simplicity.
+
+### Breaking Changes
+* `PreferenceKeysConfig` now becomes `PreferencesConfig`.
+* `PreferencesConfig` no longer supports `stringResName` since it is getting more complex to develop.
+
 ## 3.3.0 (2019-11-08)
 
 ### Enhancements